Remove templates and use concrete classes instead
Needed runtime polymorphism to allow switching memory controller (bus)
used during runtime, but this was not possible with templates.

possible solutions were to:
+ keep using templates and just use a pure abstrac class(interface) as
bus template type in cpu and then pass bus reference at runtime.
+ remove templates and use abstract classes instead

I decided not to go with either of the above two solutions and instead
use concrete classes instead to too keep things simple.

what led me to make these decision was the replies on this reddit post
[https://www.reddit.com/r/cpp/comments/1int2e5/should_i_use_templates_or_virtual_classes_for/],
which suggested that what I am trying are sign of YAGNI ( You aren't
going to need it)