Remove templates completely, Update code structure, Update tests
+ Avoid use of templated class and functions, as this increases
complexity and compile time.
+ Using templates also lead to compilation issues on windows if trying
to separate the defintion in separate files. This is issue with code in
branch feature/add-isa-impl.
+ Remove decoder as with the current approach We would have ended up
modifiying the CpuState both in decoder and executor, instead just have
a single Executor class to both decode and execute the instructions.
This leads to a single place where CpuState is ever modified
+ Refactor tests